pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "snapgrab"
version = "0.1.0"
description = "URL to screenshot with metadata. Python MCP server included. Claude Vision optimized."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "QuartzUnit" }]
keywords = ["screenshot", "capture", "web", "mcp", "playwright", "llm", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
]
dependencies = [
    "playwright>=1.49",
    "Pillow>=10.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
mcp = ["fastmcp>=2.0"]
all = ["snapgrab[mcp]"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.9"]

[project.scripts]
snapgrab = "snapgrab.__main__:main"
snapgrab-mcp = "snapgrab.mcp_server:main"

[project.urls]
Homepage = "https://github.com/QuartzUnit/snapgrab"
Repository = "https://github.com/QuartzUnit/snapgrab"
Issues = "https://github.com/QuartzUnit/snapgrab/issues"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
